added unit tests for cartcontroller

+++ src/main/java/com/darglk/onlineshop/controller/CartController.java
@@ -8,6 +8,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.RequestAttribute;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
 import org.springframework.web.bind.annotation.RequestParam;
@@ -24,28 +25,34 @@ public class CartController {
 	
 	@RequestMapping(value="show", method=RequestMethod.GET)
 	public String showCart(Model model, HttpServletRequest httpRequest) {
-		Cart cart = cartService.findCart((Long)httpRequest.getSession().getAttribute("cart_id"));
+		Cart cart = cartService.findCart(findCartIdInSession(httpRequest));
+		if(cart == null) {
+			return "redirect:/";
+		}
 		model.addAttribute("cart", cart);
 		model.addAttribute("totalPrice", cart.totalPrice().setScale(2, RoundingMode.HALF_UP));
 		return "shoppingCart";
 	}
 	
 	@RequestMapping(value="update", method=RequestMethod.POST)
-	public String updateCart(Model model, HttpServletRequest httpRequest, @RequestParam("product_ids[]") Long[] productIds,
-			@RequestParam("quantity[]") Long[] quantity) {
+	public String updateCart(Model model, HttpServletRequest httpRequest, @RequestAttribute("product_ids[]") Long[] productIds,
+			@RequestAttribute("quantity[]") Long[] quantity) {
 		
-		Long cartId = (Long)httpRequest.getSession().getAttribute("cart_id");
+		Long cartId = findCartIdInSession(httpRequest);
+		if(checkIfUpdateCartParamsAreEmptyOrNull(productIds, quantity) || (cartId == null)) {
+			return "redirect:/cart/show";
+		}
 		Cart cart = cartService.updateProductQuantity(cartId, productIds, quantity);
 		
 		model.addAttribute("cart", cart);
 		model.addAttribute("totalPrice", cart.totalPrice().setScale(2, RoundingMode.HALF_UP));
 		return "shoppingCart";
 	}
-	
+
 	@RequestMapping(value="removeItem/{id}", method=RequestMethod.GET)
 	public String removeItemFromCart(Model model, HttpServletRequest httpRequest, @PathVariable("id") Long id) {
-		Cart cart = cartService.findCart((Long)httpRequest.getSession().getAttribute("cart_id"));
-		if(cart != null) {
+		Cart cart = cartService.findCart(findCartIdInSession(httpRequest));
+		if((cart != null) && (id != null)) {
 			cartService.removeItemFromCart(id);
 		}
 		return "redirect:/cart/show";
@@ -53,21 +60,43 @@ public class CartController {
 	
 	@RequestMapping(value="clearCart", method=RequestMethod.GET)
 	public String clearCart(Model model, HttpServletRequest httpRequest) {
-		Long cartId = (Long)httpRequest.getSession().getAttribute("cart_id");
-		cartService.clearCart(cartId);
+		Long cartId = findCartIdInSession(httpRequest);
+		if(cartId != null) {
+			cartService.clearCart(cartId);
+		}
 		return "redirect:/product/list";
 	}
 	
 	@RequestMapping(value = "/add_to_cart", method = RequestMethod.POST)
-	public String addToCart(@RequestParam("product_id") Long id, Model model, HttpServletRequest httpRequest) {
+	public String addToCart(@RequestAttribute("product_id") Long id, Model model, HttpServletRequest httpRequest) {
 		Cart cart = null;
 		if(httpRequest.getSession().getAttribute("cart_id") == null) {
 			cart = cartService.createCart();
 			httpRequest.getSession().setAttribute("cart_id", cart.getId());
 		} else {
-			cart = cartService.findCart((Long)httpRequest.getSession().getAttribute("cart_id"));
+			cart = cartService.findCart(findCartIdInSession(httpRequest));
 		}
-		cart = cartService.addItemToCart(cart.getId(), id);
+		
+		if(id != null) {
+			cart = cartService.addItemToCart(cart.getId(), id);
+		}
 		return "redirect:/cart/show";
+	}
+	
+	private Long findCartIdInSession(HttpServletRequest httpRequest) {
+		return (Long)httpRequest.getSession().getAttribute("cart_id");
+	}
+	
+	private boolean checkIfUpdateCartParamsAreEmptyOrNull(Long[] productIds, Long[] quantity) {
+		return checkIfProductIdsIsNullOrEmpty(productIds) || 
+				checkIfProductQuantityIsNullOrEmpty(quantity);
+	}
+	
+	private boolean checkIfProductIdsIsNullOrEmpty(Long[] productIds) {
+		return productIds == null || (productIds.length == 0);
+	}
+	
+	private boolean checkIfProductQuantityIsNullOrEmpty(Long[] quantity) {
+		return quantity == null || (quantity.length == 0);
 	}
 }
